requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
biopython=1.81=pypi_0
bzip2=1.0.8=h3422bc3_4
ca-certificates=2023.7.22=hf0a4a13_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.2=pypi_0
distro=1.8.0=pypi_0
exceptiongroup=1.1.3=pypi_0
filelock=3.13.1=pypi_0
fsspec=2023.10.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.1=pypi_0
httpx=0.25.1=pypi_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=pypi_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.44.0=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
ncurses=6.4=h463b476_2
numpy=1.26.1=pypi_0
openssl=3.1.4=h0d3ecfb_0
packaging=23.2=pypi_0
pip=23.1.2=pyhd8ed1ab_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pymed=0.8.9=pypi_0
python=3.10.13=h2469fbe_0_cpython
python-dotenv=1.0.0=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h92ec313_1
requests=2.31.0=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
sniffio=1.3.0=pypi_0
tk=8.6.13=hb31c410_0
tokenizers=0.14.1=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.8.0=pypi_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.7=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
anthropic~=0.5.0